Put SVGLayoutElement in its own file, layouting for pacman plugs
